requirements.txt
airtable-python-wrapper (0.8.0)
astroid (2.2.5)
beautifulsoup4 (4.8.0)
certifi (2018.1.18)
chardet (3.0.4)
cssselect (1.0.3)
Django (2.0)
idna (2.6)
isort (4.3.21)
lazy-object-proxy (1.4.1)
mccabe (0.6.1)
pip (9.0.1)
pkg-resources (0.0.0)
pylint (2.3.1)
python-decouple (3.1)
pytz (2017.3)
requests (2.18.4)
setuptools (39.0.1)
six (1.11.0)
soupsieve (1.9.2)
typed-ast (1.4.0)
urllib3 (1.22)
wrapt (1.11.2)
